# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LOGIK GAME API",
        "description": "API backend du jeu télévisé LOGIK GAME — gestion des sessions, manches, questions, joueurs, temps réel et projection.",
        "contact": {
            "name": "NerdX",
            "email": "contact@nerdx.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/admin/sessions": {
            "get": {
                "tags": [
                    "Sessions"
                ],
                "summary": "Lister les sessions",
                "description": "Retourne toutes les sessions de l'admin connecté.",
                "operationId": "38b840f71f700e9159c5ebb3c89829a0",
                "responses": {
                    "200": {
                        "description": "Liste des sessions"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Sessions"
                ],
                "summary": "Créer une session",
                "description": "Crée une nouvelle session de jeu avec 8 manches par défaut.",
                "operationId": "8581b931c53e63a0ce83254db952b9c7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "scheduled_at",
                                    "max_players"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "LOGIK S1E01"
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "scheduled_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "max_players": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "cover_image": {
                                        "description": "Image de couverture (max 5Mo)",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "registration_opens_at": {
                                        "description": "Début des inscriptions",
                                        "type": "string",
                                        "format": "date-time",
                                        "nullable": true
                                    },
                                    "registration_closes_at": {
                                        "description": "Fin des inscriptions (doit être après registration_opens_at)",
                                        "type": "string",
                                        "format": "date-time",
                                        "nullable": true
                                    },
                                    "preselection_opens_at": {
                                        "description": "Début de la pré-sélection",
                                        "type": "string",
                                        "format": "date-time",
                                        "nullable": true
                                    },
                                    "preselection_closes_at": {
                                        "description": "Fin de la pré-sélection (doit être après preselection_opens_at)",
                                        "type": "string",
                                        "format": "date-time",
                                        "nullable": true
                                    },
                                    "reconnection_delay": {
                                        "description": "Délai de reconnexion en secondes (5-120)",
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Session créée"
                    },
                    "422": {
                        "description": "Validation échouée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/sessions/{session}": {
            "get": {
                "tags": [
                    "Sessions"
                ],
                "summary": "Détail d'une session",
                "description": "Retourne le détail complet d'une session avec ses manches.",
                "operationId": "5c0159bac2ca6d55b2edd435bc5bf01c",
                "parameters": [
                    {
                        "name": "session",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail session"
                    },
                    "404": {
                        "description": "Session introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Sessions"
                ],
                "summary": "Modifier une session",
                "description": "Met à jour une session (statut Draft ou RegistrationOpen uniquement).",
                "operationId": "5e677d0895bd5bd432b8ccf6174955d0",
                "parameters": [
                    {
                        "name": "session",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "scheduled_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "max_players": {
                                        "type": "integer"
                                    },
                                    "cover_image": {
                                        "description": "Image de couverture (max 5Mo)",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "remove_cover_image": {
                                        "description": "Supprimer l'image de couverture existante",
                                        "type": "boolean",
                                        "nullable": true
                                    },
                                    "registration_opens_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "nullable": true
                                    },
                                    "registration_closes_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "nullable": true
                                    },
                                    "preselection_opens_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "nullable": true
                                    },
                                    "preselection_closes_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "nullable": true
                                    },
                                    "reconnection_delay": {
                                        "description": "Délai de reconnexion en secondes (5-120)",
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Session modifiée"
                    },
                    "422": {
                        "description": "Statut invalide"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Sessions"
                ],
                "summary": "Supprimer une session",
                "description": "Supprime une session (statut Draft uniquement).",
                "operationId": "aec3530c8485962386d6be00bbe6fd74",
                "parameters": [
                    {
                        "name": "session",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Session supprimée"
                    },
                    "422": {
                        "description": "Statut invalide"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token Sanctum admin (obtenu via POST /admin/login)",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            },
            "playerToken": {
                "type": "apiKey",
                "description": "Token d'accès joueur (64 caractères, reçu après sélection)",
                "name": "X-Player-Token",
                "in": "header"
            }
        }
    },
...
}